+++ include/dev_lib/boost/delay_deadline_timer.h
@@ -7,7 +7,7 @@ class DelayDeadlineTimer : public DeadlineTimer
 {
 public:
   typedef DeadlineTimer::finish_func_type finish_func_type;
-  DelayDeadlineTimer(boost::asio::io_service& io, double sleepSeconds, finish_func_type callbackFinish) : DeadlineTimer(io, sleepSeconds, callbackFinish) { }
+  DelayDeadlineTimer(boost::asio::io_service& io, float sleepSeconds, finish_func_type callbackFinish) : DeadlineTimer(io, sleepSeconds, callbackFinish) { }
   
 private:
   virtual void callback(const boost::system::error_code &err);
